+++ CHANGES.md
@@ -8,6 +8,7 @@ Revision history for `Afterimage`, a Commodore 64 graphics library with a built-
 
 * Bug fixed: do not throw `NullPointerException: Cannot invoke "java.net.URL.getProtocol()" because "url" is null` exception when loading `Palette` data `fromJson` string in case package JAR resources could not have been successfully retrieved
 * Enhancement: report malformed JSON input when loading `Palette` data `fromJson` string only if it has been truly malformed (throw all other exceptions in a verbatim mode)
+* Bug fixed: do not throw `FileSystemAlreadyExistsException` exception when instantiating a new `Palette` object from an internal JAR resource path
 * `Scala` version upgraded to 3.7.3
 * `ScalaTest` version upgraded to 3.2.19
 * `sbt` version upgraded to 1.11.6